+++ src/middlewares/types.ts
@@ -2,4 +2,4 @@ import {RequestHandler} from 'express';
 
 export type AuthenticatedRequestMiddleware = () => RequestHandler;
 export type EnsureInstalledMiddleware = () => RequestHandler;
-export type EnsureCSPMiddleware = () => RequestHandler;
+export type CspHeadersMiddleware = () => RequestHandler;
