+++ src/types.ts
@@ -11,7 +11,7 @@ import {AuthConfigInterface, AuthMiddleware} from './auth/types';
 import {WebhooksConfigInterface, WebhooksMiddleware} from './webhooks/types';
 import {
   AuthenticatedRequestMiddleware,
-  EnsureCSPMiddleware,
+  CspHeadersMiddleware,
   EnsureInstalledMiddleware,
 } from './middlewares/types';
 
@@ -50,8 +50,8 @@ export interface ShopifyApp<
   api: Shopify<R, S>;
   auth: AuthMiddleware;
   authenticatedRequest: AuthenticatedRequestMiddleware;
+  cspHeaders: CspHeadersMiddleware;
   webhooks: WebhooksMiddleware;
-  ensureCSP: EnsureCSPMiddleware;
   ensureInstalled: EnsureInstalledMiddleware;
 }
